+++ internal/provider/schema/job_definition/databricks_input_option.go
@@ -1,0 +1,86 @@
+package job_definition
+
+import (
+	planModifier "terraform-provider-trocco/internal/provider/planmodifier"
+
+	"github.com/hashicorp/terraform-plugin-framework-validators/int64validator"
+	"github.com/hashicorp/terraform-plugin-framework-validators/listvalidator"
+	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
+)
+
+func DatabricksInputOptionSchema() schema.Attribute {
+	return schema.SingleNestedAttribute{
+		Optional:            true,
+		MarkdownDescription: "Attributes of source databricks",
+		Attributes: map[string]schema.Attribute{
+			"databricks_connection_id": schema.Int64Attribute{
+				Required: true,
+				Validators: []validator.Int64{
+					int64validator.AtLeast(1),
+				},
+				MarkdownDescription: "ID of Databricks connection",
+			},
+			"catalog_name": schema.StringAttribute{
+				Required:            true,
+				MarkdownDescription: "Catalog name",
+				Validators: []validator.String{
+					stringvalidator.UTF8LengthAtLeast(1),
+				},
+			},
+			"schema_name": schema.StringAttribute{
+				Required:            true,
+				MarkdownDescription: "Schema name",
+				Validators: []validator.String{
+					stringvalidator.UTF8LengthAtLeast(1),
+				},
+			},
+			"query": schema.StringAttribute{
+				Required:            true,
+				MarkdownDescription: "SQL query to execute",
+				Validators: []validator.String{
+					stringvalidator.UTF8LengthAtLeast(1),
+				},
+			},
+			"fetch_rows": schema.Int64Attribute{
+				Optional:            true,
+				Computed:            true,
+				MarkdownDescription: "Number of records processed by the cursor at one time",
+				Validators: []validator.Int64{
+					int64validator.AtLeast(1),
+				},
+			},
+			"input_option_columns": schema.ListNestedAttribute{
+				Required:            true,
+				MarkdownDescription: "List of columns to be retrieved and their types",
+				NestedObject: schema.NestedAttributeObject{
+					Attributes: map[string]schema.Attribute{
+						"name": schema.StringAttribute{
+							Required: true,
+							Validators: []validator.String{
+								stringvalidator.UTF8LengthAtLeast(1),
+							},
+							MarkdownDescription: "Column name",
+						},
+						"type": schema.StringAttribute{
+							Required:            true,
+							MarkdownDescription: "Column type",
+							Validators: []validator.String{
+								stringvalidator.OneOf("boolean", "long", "timestamp", "double", "string", "json"),
+							},
+						},
+					},
+				},
+				Validators: []validator.List{
+					listvalidator.SizeAtLeast(1),
+				},
+			},
+			"custom_variable_settings": CustomVariableSettingsSchema(),
+		},
+		PlanModifiers: []planmodifier.Object{
+			&planModifier.DatabricksInputOptionPlanModifier{},
+		},
+	}
+}