+++ internal/provider/schema/job_definition/input_option.go
@@ -24,6 +24,7 @@ func InputOptionSchema() schema.Attribute {
 			"kintone_input_option":             KintoneInputOptionSchema(),
 			"yahoo_ads_api_yss_input_option":   YahooAdsApiYssInputOptionSchema(),
 			"hubspot_input_option":             HubspotInputOptionSchema(),
+			"databricks_input_option":          DatabricksInputOptionSchema(),
 		},
 		PlanModifiers: []planmodifier.Object{
 			&planModifier.InputOptionPlanModifier{},