+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.21.0
+FEATURES:
+- `trocco_connection` resource:
+  - Added support for `databricks` type
+- `trocco_job_definition` resource:
+  - Added support for `databricks` input option
+  - Added support for `databricks` output option
+
 ## 0.20.0
 - `trocco_user` resource:
   - Added support for `password_auto_generated` field for user invitations with auto-generated passwords